requirements.txt
CryptoPlus==1.0
pycrypto==2.6.1
